Refresh cues and displaying video

+++ src/main/java/com/mh/lamp/cue/CueRepository.java
@@ -2,6 +2,6 @@ package com.mh.lamp.cue;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CueRepository extends JpaRepository<Cue, Integer> {
+public interface CueRepository extends JpaRepository<Cue, Double> {
 
 }
